feature(analytics): add actionable analytics
