+++ imports/plugins/included/actionableAnalytics/client/components/RetailDashBoard.js
@@ -38,7 +38,7 @@ export const RetailDashBoard = (props) => {
       {/* gross profit box */}
       <Box
         boxTitle="GROSS PROFIT"
-        boxValue={formatPriceString(grossProfit)}
+        boxValue={formatPriceString(grossProfit || 0)}
       />
       {/* placed orders board */}
       <Box
